+++ Tests/StreamChatTests/Controllers/ChannelListController/ChannelListController_Tests.swift
@@ -1088,14 +1088,14 @@ final class ChannelListController_Tests: XCTestCase {
 
         // When all the channel ids are in DB.
         try assertFilterPredicate(
-            .in(.id, values: expectedCids.map(\.rawValue)),
+            .in(.cid, values: expectedCids),
             channelsInDB: expectedChannels + unexpectedChannels,
             expectedResult: expectedCids
         )
 
         // When not all the channel ids are in DB.
         try assertFilterPredicate(
-            .in(.id, values: expectedCids.map(\.rawValue)),
+            .in(.cid, values: expectedCids),
             channelsInDB: expectedChannels.dropLast() + unexpectedChannels,
             expectedResult: [cid1, cid2]
         )
@@ -1489,7 +1489,7 @@ final class ChannelListController_Tests: XCTestCase {
         )
     }
 
-    func test_filterPredicate_inWithArrayOfIds_returnsExpectedResults() throws {
+    func test_filterPredicate_inWithArrayOfCids_returnsExpectedResults() throws {
         let chatIds: [String] = [
             "suggestions-63986de56549624f314b75cb",
             "suggestions-6ukh3986de56549624f314b75cjkhagfdkjhab",
@@ -1500,7 +1500,7 @@ final class ChannelListController_Tests: XCTestCase {
         let channelIds = chatIds.map { ChannelId(type: .custom("daisy-dashboard"), id: $0) }
 
         try assertFilterPredicate(
-            .in(.id, values: channelIds.map(\.rawValue)),
+            .in(.cid, values: channelIds),
             channelsInDB: [
                 .dummy(channel: .dummy(cid: channelIds[0])),
                 .dummy(channel: .dummy(cid: channelIds[1])),
@@ -1700,6 +1700,23 @@ final class ChannelListController_Tests: XCTestCase {
                 .dummy(channel: .dummy(team: .unique))
             ],
             expectedResult: [cid]
+        )
+    }
+
+    func test_filterPredicate_id_returnsExpectedResults() throws {
+        let cid1 = ChannelId(type: .commerce, id: "123")
+        let cid2 = ChannelId(type: .livestream, id: "123")
+
+        try assertFilterPredicate(
+            .equal(.id, to: "123"),
+            channelsInDB: [
+                .dummy(channel: .dummy(cid: cid1)),
+                .dummy(channel: .dummy(cid: .init(type: .commerce, id: "123123"))),
+                .dummy(channel: .dummy()),
+                .dummy(channel: .dummy()),
+                .dummy(channel: .dummy(cid: cid2))
+            ],
+            expectedResult: [cid1, cid2]
         )
     }