+++ CHANGELOG.md
@@ -3,6 +3,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
+### 🔄 Changed
+
+# [4.78.0](https://github.com/GetStream/stream-chat-swift/releases/tag/4.78.0)
+_April 24, 2025_
+
 ## StreamChat
 ### ✅ Added
 - Add `CurrentUserUnreads.totalUnreadMessagesCount` [#3651](https://github.com/GetStream/stream-chat-swift/pull/3651)